requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
flask-socketio==5.3.2
python-dotenv==0.19.0
gevent==21.12.0
flask-bcrypt
flask-migrate
